pyproject.toml
[project]
name = "emb"
dynamic = ["version"]
description = "Embed, index, and search text corpora with sentence-transformers"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
    "typer>=0.19.2",
    "rich>=14.1.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
image = [
    "open-clip-torch>=2.24.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
]
rerank = [
    "sentence-transformers>=2.2.0",
]
cag = [
    "google-genai>=1.0.0",
]
contextualize = [
    "google-genai>=1.0.0",
]
dev = [
    "pytest>=9.0.1",
]

[project.scripts]
emb = "emb.cli:app"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.targets.wheel]
packages = ["src/emb"]

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
fallback-version = "0.0.0"
